requirements.txt
pdfplumber==0.5.28
openpyxl==3.0.9
pandas==1.3.3
python-docx==0.8.11
google.generativeai